Add maxCombinedHeaderSize option to truncate logs if HTTP header gets too big
The new option can be set programmatically via $firePhp->setOption() or via
an HTTP header in the request (typically sent by the browser extension).

The limit is expressed in kilobytes. FirePHP always reserve 10kb for
non-FirePHP headers. So if the overall limit for all headers for your browser
is 256kb, then send the header "X-FirePHP-MaxCombinedHeaderSize: 256"